+++ lib/xml/XmldocFileParser.ts
@@ -30,15 +30,16 @@ import {
  * FileParser implementation that uses xmldoc library.
  * Preserves and exposes positions.
  */
-export class XmldocFileParser implements FileParser<XmlDocTreeNode> {
+export class XmldocFileParser implements FileParser<XmldocTreeNode> {
 
     public rootName = "xml";
 
-    public async toAst(f: ProjectFile): Promise<XmlDocTreeNode> {
+    public async toAst(f: ProjectFile): Promise<XmldocTreeNode> {
         try {
-            const document = new XmlDocument(await f.getContent());
+            const content = await f.getContent();
+            const document = new XmlDocument(content);
             // console.log("DOC is " + JSON.stringify(document));
-            return new XmldocTreeNodeImpl(document, undefined);
+            return new XmldocTreeNodeImpl(document, undefined, content);
         } catch (err) {
             logger.warn("Could not parse XML document at '%s'", f.path, err);
             return undefined;
@@ -50,26 +51,33 @@ export class XmldocFileParser implements FileParser<XmlDocTreeNode> {
 /**
  * Allows further operations specific to XML elements
  */
-export interface XmlDocTreeNode extends TreeNode {
+export interface XmldocTreeNode extends TreeNode {
 
     /**
      * Value inside the element: Not the same as it's value
      */
     innerValue: string;
+
+    /**
+     * Specialize return type
+     * @return {XmldocTreeNode[]}
+     */
+    $children: XmldocTreeNode[];
+
 }
 
-export function isXmlDocTreeNode(tn: TreeNode): tn is XmlDocTreeNode {
-    const maybe = tn as XmlDocTreeNode;
+export function isXmldocTreeNode(tn: TreeNode): tn is XmldocTreeNode {
+    const maybe = tn as XmldocTreeNode;
     return !!maybe.innerValue;
 }
 
-class XmldocTreeNodeImpl implements XmlDocTreeNode {
+class XmldocTreeNodeImpl implements XmldocTreeNode {
 
-    public get $children(): TreeNode[] {
+    public get $children(): XmldocTreeNode[] {
         return this.xd.children
             .filter(kid => kid.type === "element")
             .map(k =>
-                new XmldocTreeNodeImpl(k as XmlElement, this));
+                new XmldocTreeNodeImpl(k as XmlElement, this, this.rawDoc));
     }
 
     public get $name(): string {
@@ -85,7 +93,25 @@ class XmldocTreeNodeImpl implements XmlDocTreeNode {
      * @return {string}
      */
     public get $value(): string {
-        return this.xd.toString({ preserveWhitespace: true, compressed: false, trimmed: false });
+        // toString may not be accurate, as per xmldoc readme, but we can work with it
+        // as the offset will be accurate
+        const fromXmldocToString = this.xd.toString({ preserveWhitespace: true, compressed: false, trimmed: false });
+        const fromRawDoc = this.rawDoc.substr(this.$offset, fromXmldocToString.length);
+        if (fromRawDoc !== fromXmldocToString) {
+            // In this case, check we have all the non whitespace characters from the toString value
+            const nonWhitespaceCount = fromXmldocToString.replace(/\s+/g, "").length;
+            let included = 0;
+            let str = "";
+            for (let i = 0; i < fromXmldocToString.length && included < nonWhitespaceCount; i++) {
+                const c = this.rawDoc.substr(this.$offset).charAt(i);
+                if (!["\n", " ", "\t", "\r"].includes(c)) {
+                    ++included;
+                }
+                str += c;
+            }
+            return str;
+        }
+        return fromXmldocToString;
     }
 
     public get innerValue(): string {
@@ -93,7 +119,8 @@ class XmldocTreeNodeImpl implements XmlDocTreeNode {
     }
 
     constructor(private readonly xd: XmlElement,
-                public readonly $parent: TreeNode) {
+                public readonly $parent: TreeNode,
+                private readonly rawDoc: string) {
         // Add attributes to this
         for (const propName of Object.getOwnPropertyNames(this.xd.attr)) {
             (this as any)[propName] = this.xd.attr[propName];